+++ lib/src/analyzers/lint_analyzer/rules/rules_list/check_for_equals_in_render_object_setters/visitor.dart
@@ -97,6 +97,7 @@ class _ReturnVisitor extends RecursiveAstVisitor<void> {
 
     final type = node.expression?.staticType;
 
+    // ignore: deprecated_member_use
     if (type == null || type.isVoid) {
       _hasValidReturn = true;
     }