+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## Unreleased
 
 * fix: handle dynamics in map literals for [`avoid-dynamic`](https://dcm.dev/docs/individuals/rules/common/avoid-dynamic).
+* fix: change anti-patterns default severity to `warning`.
+* feat: add static code diagnostic [`prefer-define-hero-tag`](https://dcm.dev/docs/individuals/rules/common/prefer-define-hero-tag).
+* chore: restrict `analyzer` version to `>=5.1.0 <5.8.0`.
+* feat: add static code diagnostic [`avoid-substring`](https://dcm.dev/docs/individuals/rules/common/avoid-substring).
 
 ## 5.6.0